% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_palette.R
\name{colorblind_friendly}
\alias{colorblind_friendly}
\title{Colorblind-Friendly Palette Check}
\usage{
colorblind_friendly(palette_name, type = "all")
}
\arguments{
\item{palette_name}{Name of Palette. Choices are: \code{Beija_Flor}, \code{Grande_Rio},
\code{Imperatriz_Leopoldinense}, \code{Imperio_Serrano}, \code{Mangueira}, \code{Padre_Miguel},
\code{Paraiso_Tuiuti}, \code{Portela}, \code{Salgueiro}, \code{Sao_Clemente}, \code{Tijuca},
\code{Uniao_Ilha}, \code{Vila_Isabel}, \code{Viradouro}.}

\item{type}{Which kind of colorblindness to check. Deuteranopia, protanopia,
and/or tritanopia}
}
\value{
TRUE/FALSE value whether palette is colorblind-friendly
}
\description{
Checks whether a palette is colorblind-friendly. Colorblind-friendliness
tested using the {colorblindcheck} package. It's possible to check if a
palette is friendly towards deuteranopia, protanopia, and/or tritanopia.
}
\examples{
colorblind_friendly("Imperio_Serrano", type = "tritanopia")
}
