% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_notes}
\alias{get_notes}
\title{Obtain notes on the evaluation process}
\usage{
get_notes(years)
}
\arguments{
\item{years}{\strong{numeric:} which years of notes to show. almost all years
since 1968 are available.}
}
\value{
A tibble. Its columns are \code{info} (notes) and \code{year}
  (parade year). Each line corresponds to all notes made in a given year.
}
\description{
\code{get_notes()} returns a tibble of notes about the evaluation process of
each year competition. They may explain such things as: changes in the rules,
substitution of judges or absence of demotion of schools.
}
\examples{
# Prints a table that shows the notes on the evaluation process of 1970 and 1980
get_notes(years = c(1970, 1980))
}
