% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_palette.R
\name{rio.paletter}
\alias{rio.paletter}
\title{Generate Samba Schools palettes}
\usage{
rio.paletter(
  palette_name,
  n,
  type = c("discrete", "continuous"),
  direction = c(1, -1),
  override.order = FALSE
)
}
\arguments{
\item{n}{Number of desired colors. If number of requested colors is beyond
the scope of the palette, colors are automatically interpolated. If n is
not provided, the length of the palette is used.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want to
automatically interpolate between colors.}

\item{direction}{Sets order of colors. Default palette is 1. If direction is
-1, palette color order is reversed}

\item{override.order}{Colors are picked from palette to maximize readability
and aesthetics. This means that colors are not always selected in
sequential order from the full palette. If override.order is set to TRUE,
colors are selected in sequential order from the full palette instead.
Default is FALSE.}

\item{palette}{Name of Palette. Choices are: \code{Beija_Flor},
\code{Grande_Rio}, \code{Imperatriz_Leopoldinense}, \code{Imperio_Serrano},
\code{Mangueira}, \code{Padre_Miguel}, \code{Paraiso_Tuiuti},
\code{Portela}, \code{Salgueiro}, \code{Sao_Clemente}, \code{Tijuca},
\code{Uniao_Ilha}, \code{Vila_Isabel}, \code{Viradouro}.}
}
\value{
A vector of colors.
}
\description{
Color palettes inspired by the shields of Carioca Samba Schools. Complete
list of palette colors and the shields that inspired them can be found
\href{https://github.com/IcaroBernardes/carnaval}{on Github}. Use
\code{\link{colorblind_friendly()}} to check whether palettes are
colorblind-friendly.
}
\examples{
rio.paletter("Beija_Flor")

rio.paletter("Grande_Rio", direction=-1)

rio.paletter("Sao_Clemente", 4, override.order=TRUE)

library(ggplot2)
ggplot(data=iris, aes(x=Species, y=Petal.Length, fill=Species)) +
geom_violin() +
scale_fill_manual(values=rio.paletter("Imperio_Serrano", 3))

ggplot(data=iris, aes(x=Sepal.Length, y=Sepal.Width, color=Species)) +
geom_point(size=2) +
scale_color_manual(values=rio.paletter("Mangueira", 3))

ggplot(data=iris, aes(x=Species, y=Sepal.Width, color=Sepal.Width)) +
geom_point(size=3) +
scale_color_gradientn(colors=rio.paletter("Padre_Miguel"))
}
\keyword{colors}
