% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_palette.R
\docType{data}
\name{RioPalettes}
\alias{RioPalettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 14.
}
\usage{
RioPalettes
}
\description{
Use names(RioPalettes) to return all possible palette names. Current choices
are: \code{Beija_Flor}, \code{Grande_Rio}, \code{Imperatriz_Leopoldinense},
\code{Imperio_Serrano}, \code{Mangueira}, \code{Padre_Miguel},
\code{Paraiso_Tuiuti}, \code{Portela}, \code{Salgueiro}, \code{Sao_Clemente},
\code{Tijuca}, \code{Uniao_Ilha}, \code{Vila_Isabel}, \code{Viradouro}. Use
\code{\link{rio.paletter}} to construct palettes.
}
\keyword{datasets}
