% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_palette.R
\name{scale_color_rio_c}
\alias{scale_color_rio_c}
\title{RioPaletter palettes for plotting with ggplot2}
\usage{
scale_color_rio_c(palette_name, direction = 1, ...)
}
\arguments{
\item{palette_name}{Name of Palette. Choices are: \code{Beija_Flor},
\code{Grande_Rio}, \code{Imperatriz_Leopoldinense}, \code{Imperio_Serrano},
\code{Mangueira}, \code{Padre_Miguel}, \code{Paraiso_Tuiuti},
\code{Portela}, \code{Salgueiro}, \code{Sao_Clemente}, \code{Tijuca},
\code{Uniao_Ilha}, \code{Vila_Isabel}, \code{Viradouro}.}

\item{direction}{Sets order of colors. Default palette is 1. If direction is
-1, palette color order is reversed}

\item{...}{Other arguments passed on to
\code{\link[ggplot2]{scale_color_gradientn}}}
}
\description{
Function for using \code{RioPaletter} colors schemes in \code{ggplot2}. Use
\code{\link{scale_color_rio_d}} and \code{\link{scale_fill_rio_d}} for
discrete scales and \code{\link{scale_color_rio_c}} and
\code{\link{scale_fill_rio_c}} for continuous scales.
}
\examples{
library(ggplot2)
ggplot(data=iris, aes(x=Sepal.Length, y=Sepal.Width, color=Sepal.Length)) +
geom_point() +
scale_color_rio_c("Salgueiro", direction=-1)
}
