% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_scores}
\alias{get_scores}
\title{Obtain scores of samba schools by year, school and/or criterion}
\usage{
get_scores(years, schools = "all", criterions = "all")
}
\arguments{
\item{years}{\strong{numeric:} which years of parade to show. almost all years since 1968 are available.}

\item{schools}{\strong{character:} which samba schools to show. Case
insensitive.}

\item{criterions}{\strong{character:} which evaluation criteria to show. Case
insensitive.}
}
\value{
A tibble. Its columns are \code{school} (samba school name),
  \code{judge_number} (judge id), \code{score} (samba school score),
  \code{criteria} (evaluation criteria), \code{judge_name} (judge name), and
  \code{year} (parade year). Each line corresponds to a score given by a
  judge to a samba school on a criteria in a given year.
}
\description{
\code{get_scores()} returns a tibble of scores of samba schools. It can be
filtered by \code{year}, \code{school} and/or \code{criterion}
}
\examples{
# Prints a table that shows the scores of Portela and
# Est치cio de S치 on the parades of 1968 and 1970 on all criterions
get_scores(years = c(1968, 1970), schools = c("portela", "Est치cio de S치"))
}
