% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_palette.R
\name{scale_colour_rio_d}
\alias{scale_colour_rio_d}
\title{RioPaletter palettes for plotting with ggplot2}
\usage{
scale_colour_rio_d(palette_name, direction = 1, override.order = FALSE, ...)
}
\arguments{
\item{palette_name}{Name of Palette. Choices are: \code{Beija_Flor}, \code{Grande_Rio},
\code{Imperatriz_Leopoldinense}, \code{Imperio_Serrano}, \code{Mangueira}, \code{Padre_Miguel},
\code{Paraiso_Tuiuti}, \code{Portela}, \code{Salgueiro}, \code{Sao_Clemente}, \code{Tijuca},
\code{Uniao_Ilha}, \code{Vila_Isabel}, \code{Viradouro}.}

\item{direction}{Sets order of colors. Default palette is 1. If direction is
-1, palette color order is reversed}

\item{override.order}{Colors are picked from palette to maximize readability.
This means that colors are not always selected in sequential order from the
full palette. If override.order is set to TRUE, colors are selected in
sequential order from the full palette instead. Default is FALSE.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}}
}
\description{
Function for using \link{RioPalettes} colors schemes in \code{ggplot2}. Use
\code{\link[=scale_color_rio_d]{scale_color_rio_d()}} and \code{\link[=scale_fill_rio_d]{scale_fill_rio_d()}} for discrete scales and
\code{\link[=scale_color_rio_c]{scale_color_rio_c()}} and \code{\link[=scale_fill_rio_c]{scale_fill_rio_c()}} for continuous scales.
}
\examples{
# Gets the scores of Mangueira from 1968 until 2018 on the Bateria criterion.
# Throws some warnings to inform the user which years are absent of the data
# (either from missing info or because the school didn't compete in the main league)
df <- get_scores(years = 1968:2018,
                 schools = "Estação primeira de Mangueira",
                 criterions = "BATERIA")

# Loads ggplot2 and dplyr
library(ggplot2)
library(dplyr)

# Creates a plot and applies the palette of Mangueira
df |>
  mutate(era = case_when(
      between(year, 1968, 1984) ~ "1968-1984",
      between(year, 1985, 2001) ~ "1985-2001",
      between(year, 2002, 2018) ~ "2002-2018"
  )) |>
  ggplot(aes(x = era, y = score, color = era)) +
    geom_jitter() +
    scale_colour_rio_d("Mangueira")
}
